requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cu128

absl-py==2.3.1
    # via rouge-score
accelerate==1.9.0
    # via
    #   -r requirements.in
    #   peft
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
attrs==25.3.0
    # via aiohttp
bitsandbytes==0.46.1
    # via -r requirements.in
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via nltk
colorama==0.4.6
    # via sacrebleu
datasets==4.0.0
    # via evaluate
datasketch==1.6.5
    # via -r requirements.in
detect-secrets==1.5.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   evaluate
    #   multiprocess
evaluate==0.4.5
    # via -r requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   torch
hf-xet==1.1.5
    # via huggingface-hub
huggingface-hub==0.33.4
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   evaluate
    #   peft
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via nltk
lxml==6.0.1
    # via sacrebleu
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   datasets
    #   evaluate
networkx==3.5
    # via torch
nltk==3.9.1
    # via
    #   -r requirements.in
    #   rouge-score
numpy==2.3.1
    # via
    #   accelerate
    #   bitsandbytes
    #   datasets
    #   datasketch
    #   evaluate
    #   pandas
    #   peft
    #   rouge-score
    #   sacrebleu
    #   scipy
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
packaging==25.0
    # via
    #   accelerate
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   peft
    #   transformers
pandas==2.3.1
    # via
    #   datasets
    #   evaluate
peft==0.16.0
    # via -r requirements.in
pillow==11.3.0
    # via torchvision
portalocker==3.2.0
    # via sacrebleu
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via
    #   accelerate
    #   peft
pyarrow==21.0.0
    # via datasets
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   detect-secrets
    #   huggingface-hub
    #   peft
    #   transformers
regex==2024.11.6
    # via
    #   nltk
    #   sacrebleu
    #   transformers
requests==2.32.4
    # via
    #   datasets
    #   detect-secrets
    #   evaluate
    #   huggingface-hub
    #   transformers
rouge-score==0.1.2
    # via -r requirements.in
sacrebleu==2.5.1
    # via -r requirements.in
safetensors==0.5.3
    # via
    #   accelerate
    #   peft
    #   transformers
scipy==1.16.1
    # via datasketch
six==1.17.0
    # via
    #   python-dateutil
    #   rouge-score
sympy==1.14.0
    # via torch
tabulate==0.9.0
    # via sacrebleu
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
    #   peft
    #   torchvision
torchvision==0.22.1+cu128
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   nltk
    #   peft
    #   transformers
transformers==4.53.3
    # via
    #   -r requirements.in
    #   peft
triton==3.3.1
    # via torch
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   huggingface-hub
    #   torch
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via requests
xxhash==3.5.0
    # via
    #   datasets
    #   evaluate
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
